unpublish f-droid issues
This article was meant to be constructive rather than destructive. Unfortunately, many people, including F-Droid project members haven misunderstood my intentions. To this day, they keep using that article to attack GrapheneOS (of which I am NOT a project member) and myself. I still believe that this is useful public information and that it doesn't get "the facts wrong". However, I can understand why people may see it as inflammatory.

Therefore, I'm taking down this article, so it won't be used to harm anyone anymore. People are free of their own choices, and after all, information grants freedom. Thanks for reading!